Add files via upload
As presented at the IATUL Fall 2022 workshop in Zurich. An example of how to leverage the Bibliometrix package to find authors whose research is similar based on the Bibliographic Coupling of their publications.
TO BE DONE: The affiliation of authors is lost during this process. Filtering the output by affiliation might be done by keeping a list in parallel.